requirements.txt
# requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.24.1
openai
python-multipart==0.0.6
pillow==10.1.0